% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncensured.R
\name{compute_test_values_uncens}
\alias{compute_test_values_uncens}
\title{Computing Quantiles for tests - statistic tests}
\usage{
compute_test_values_uncens(n)
}
\arguments{
\item{n}{integer, number of observations in full sample}
}
\description{
Computing Quantiles for tests - statistic tests
}
\keyword{internal}
