% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newstatistics.R
\name{Delta_k}
\alias{Delta_k}
\title{Computing value of statistic \eqn{\Delta_k}}
\usage{
Delta_k(k, n, sort_Y_vec_dt)
}
\arguments{
\item{k}{integer, k-th observation in sorted data}

\item{n}{integer, amount of observations in data}

\item{sort_Y_vec_dt}{data.table object, two-columns table which contains value of sorted statistic \eqn{Y}
and information of observation's censoring (1 - uncensored, 0 - censored).}
}
\value{
numeric, a value of statistic \eqn{\Delta_k}
}
\description{
Statistic represents the size of jump in \eqn{G_n(T_{(k)})} and can be described by a formula
\deqn{\Delta_k = G_n(T_{(k)}) - \lim_{t \rightarrow T_{(k)}} G_n(t), \ k = 1, \ldots n} This difference can be simplified to
calculable expression and the function \code{Delta_k} used it to computes value of statistic.
}
\keyword{internal}
