% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deritatives.R
\name{first_deritative_theta_k}
\alias{first_deritative_theta_k}
\alias{first_deritative_lambda_k}
\alias{second_deritative_theta_k}
\alias{second_deritative_lambda_k}
\alias{mix_deritative_k}
\alias{grad_k}
\alias{Hessian_k}
\title{Deritatives used in Newton-Raphson Algorithm}
\usage{
first_deritative_theta_k(T_k, delta, theta_k, lambda_k)

first_deritative_lambda_k(T_k, delta, theta_k, lambda_k)

second_deritative_theta_k(T_k, delta, theta_k, lambda_k)

second_deritative_lambda_k(T_k, delta, theta_k, lambda_k)

mix_deritative_k(T_k, delta, theta_k, lambda_k)

grad_k(T_k, delta, theta_k, lambda_k)

Hessian_k(T_k, delta, theta_k, lambda_k)
}
\arguments{
\item{T_k}{numeric, k-th observation's value}

\item{delta}{vector with 0 and 1 values, information of censorship each observation
in sample}

\item{theta_k}{numeric, k-th estimatation of parameter theta (shape parameter)}

\item{lambda_k}{numeric, k-th estimatation of parameter lambda (scale parameter)}
}
\description{
Deritatives used in Newton-Raphson Algorithm
}
\keyword{internal}
