% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deritatives.R
\name{first_deritative_theta_k}
\alias{first_deritative_theta_k}
\alias{first_deritative_lambda_k}
\alias{second_deritative_theta_k}
\alias{second_deritative_lambda_k}
\alias{mix_deritative_k}
\alias{grad_k}
\alias{Hessian_k}
\title{Derivatives of loglikehood function of sample from Weibull distribution
(full and censored samples)}
\usage{
first_deritative_theta_k(T_k, delta, theta_k, lambda_k)

first_deritative_lambda_k(T_k, delta, theta_k, lambda_k)

second_deritative_theta_k(T_k, delta, theta_k, lambda_k)

second_deritative_lambda_k(T_k, delta, theta_k, lambda_k)

mix_deritative_k(T_k, delta, theta_k, lambda_k)

grad_k(T_k, delta, theta_k, lambda_k)

Hessian_k(T_k, delta, theta_k, lambda_k)
}
\arguments{
\item{T_k}{numeric, a value of k-th observation, a point where the function are computed}

\item{delta}{vector with 0 and 1 values, information of censorship all observations
in sample}

\item{theta_k}{numeric, k-th estimation of parameter \eqn{\theta} (shape parameter)}

\item{lambda_k}{numeric, k-th estimation of parameter \eqn{\lambda} (scale parameter)}
}
\description{
The loglikehood function of sample from from Weibull distribution is described by from:
\deqn{l(\lambda, \theta|\textbf{X}) = \log{(\theta)}\sum^n_{i=1}\delta_i
 - \theta\log{(\lambda)}\sum^n_{i=1}\delta_i + (\theta - 1)\sum^n_{i=1}
 \delta_i\log{(X_i)} - \lambda^{-\theta}\sum^n_{i=1}X_i^{\theta}.}
If the sample is full \eqn{\sum^n_{i=1}\delta_i = n}. The functions below compute values of
first-order and second-order partial derivatives of \eqn{l(\lambda, \theta|\textbf{X})} with respect
to \eqn{\theta}, \eqn{\lambda} at \code{T_k}. Additionally, they compute mixed partial derivative,
Hessian and gradient of \eqn{l(\lambda, \theta|\textbf{X})}.
}
\keyword{internal}
