% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{Est_kaplan_meier}
\alias{Est_kaplan_meier}
\title{Kaplan-Meier Estimator of distribution of sample}
\usage{
Est_kaplan_meier(t, n, sort_vec_dt)
}
\arguments{
\item{t}{numeric, the value, where the estimator will be counted}

\item{n}{integer, number of observations in data}

\item{sort_vec_dt}{a data table which contains value of sorted statistic and
information of observation's censoring}
}
\description{
Kaplan-Meier Estimator of distribution of sample
}
\keyword{internal}
