% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncensured.R
\name{generation_quantiles_uncens}
\alias{generation_quantiles_uncens}
\title{Generate list of data.frames with statistic values to compute Quantiles}
\usage{
generation_quantiles_uncens(m, n1, ...)
}
\arguments{
\item{m}{integer, amount of iterations of Parallel loop}

\item{n1}{integer, amount of observations in full sample}

\item{...}{integer, another amounts of observations in full sample}
}
\description{
Generate list of data.frames with statistic values to compute Quantiles
}
\keyword{internal}
