% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncensured.R
\name{getDataFrameForSample}
\alias{getDataFrameForSample}
\title{Find all statistics' value for choosen type of data and get data.frame}
\usage{
getDataFrameForSample(list_of_vals, strColName)
}
\arguments{
\item{list_of_vals}{list of objects from class data.frame; result of \link{\code{generation_quantiles_uncens}} function}

\item{strColName}{characteristic, type of data (uncensored/censored and amount of observations)}
}
\description{
Find all statistics' value for choosen type of data and get data.frame
}
\keyword{internal}
