% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncensured.R
\name{getQuantileForTest}
\alias{getQuantileForTest}
\title{Find the estimation of quantile for choosen test value and type of data}
\usage{
getQuantileForTest(df, strTestName, m = 10000, alpha = 0.05)
}
\arguments{
\item{df}{data frame object...}

\item{m}{description}

\item{alpha}{description}

\item{strColName}{characteristic, type of data (uncensored/censored and amount of observations)}
}
\description{
Find the estimation of quantile for choosen test value and type of data
}
\keyword{internal}
