% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncensured.R
\name{getQuantiles_DF}
\alias{getQuantiles_DF}
\title{Compute Quantiles for each test and each type of data}
\usage{
getQuantiles_DF(alpha, m, n1, ...)
}
\arguments{
\item{alpha}{data frame object...}

\item{m}{description}

\item{n1}{description}

\item{...}{description}
}
\description{
Compute Quantiles for each test and each type of data
}
\keyword{internal}
