% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncensured.R
\name{getQuantilesForEachTest}
\alias{getQuantilesForEachTest}
\title{Find the estimation of quantile for all test value and one type of data}
\usage{
getQuantilesForEachTest(df, m, alpha = 0.05)
}
\arguments{
\item{df}{data frame object...}

\item{m}{description}

\item{alpha}{description}
}
\description{
Find the estimation of quantile for all test value and one type of data
}
\keyword{internal}
