% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncensured.R
\name{getStatisticTestValue}
\alias{getStatisticTestValue}
\title{Find statistic's value for choosen type of data}
\usage{
getStatisticTestValue(list_of_vals, i, strColName)
}
\arguments{
\item{list_of_vals}{list of objects from class data.frame; result of \link{\code{generation_quantiles_uncens}} function}

\item{i}{integer, a position of test statistic in rownames}

\item{strColName}{characteristic, type of data (uncensored/censored and amount of observations)}
}
\description{
Find statistic's value for choosen type of data
}
\keyword{internal}
