% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistictsUsedInComparison.R
\name{KS_test}
\alias{KS_test}
\title{Computing value of Komogorov-Smirnov test (\eqn{KS_n^*}) for right-censored samples}
\source{
J.A. Koziol, S.B. Green. “A Cram ́er-von Mises statistic for randomly censored data”. Biometrika 63 (1976), pp. 465–474
}
\usage{
KS_test(Y_vec_dt, n)
}
\arguments{
\item{Y_vec_dt}{a data.table object, two-columns table which contains value of statistic \eqn{Y}
and information of observation's censoring (1 - uncensored, 0 - censored).}

\item{n}{integer, amount of observations in data}
}
\value{
numeric,a value of test statistic \eqn{KS_n^*}
}
\description{
Classic one-sample Kolmogorov-Smirnov test (\eqn{KS_n^*}) which was modified to usage for data with
presence of censoring.
}
\details{
The test can be applied for full samples as well because of a Kaplan-Meier
estimator's universality. However, it can be lead too much simplification of algorithm to computing quantiles.
Hence, in that case it is recommended use \code{\link{KS_test_full}}.
}
\keyword{internal}
