% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistictsUsedInComparison.R
\name{LS_test}
\alias{LS_test}
\title{Liao and Shimokawa test - Right-Censored Sample Case}
\usage{
LS_test(Y_vec_dt, n)
}
\arguments{
\item{Y_vec_dt}{a data table which contains value of transformed data statistic Y and
information of observation's censoring}

\item{n}{integer, number of observations in data}
}
\description{
Liao and Shimokawa test - Right-Censored Sample Case
}
\keyword{internal}
