% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistictsUsedInComparison.R
\name{LS_test_full}
\alias{LS_test_full}
\title{Computing value of Liao and Shimokawa test (\eqn{LS_n}) for full samples}
\source{
M. Liao, T. Shimokawa. “A new goodness-of-fit test for type-I extreme-value and
2-parameter weibull distributions with estimated parameters”.
Journal of Statistical Computation and Simulation 64 (1999), pp. 23–48
}
\usage{
LS_test_full(X_vec, n)
}
\arguments{
\item{X_vec}{numeric, the values of observations from uncensored sample}

\item{n}{integer, amount of observations in data}
}
\value{
numeric,a value of test statistic \eqn{LS_n}
}
\description{
Classic Liao and Shimokawa test (\eqn{LS_n}) applied to data without censoring.
}
\details{
By contrast to another tests, Liao and Shimokawa's statistic is constructed base on
the least squares estimators of parameters instead of maximum likelihood ones.
}
\keyword{internal}
