% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton.R
\name{mle_lambda_full_sample}
\alias{mle_lambda_full_sample}
\title{Starting points of Newton-Raphson Algorithm (full sample case) - parameter \eqn{\lambda}}
\usage{
mle_lambda_full_sample(
  n,
  T_vals,
  theta_hat_est = mle_theta_full_sample(n, T_vals)
)
}
\arguments{
\item{n}{integer, number of observations in data}

\item{T_vals}{vector of observations}

\item{theta_hat_est}{estimation of parameter \eqn{\theta} (shape parametr); result of
function \code{\link{mle_theta_full_sample}}}
}
\description{
Starting points of Newton-Raphson Algorithm (full sample case) - parameter \eqn{\lambda}
}
\keyword{internal}
