% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton.R
\name{mle_theta_full_sample}
\alias{mle_theta_full_sample}
\title{Starting points of Newton-Raphson Algorithm (full sample case) - parameter \eqn{\theta}}
\usage{
mle_theta_full_sample(n, T_vals)
}
\arguments{
\item{n}{integer, number of observations in data}

\item{T_vals}{vector of observations}
}
\description{
Starting points of Newton-Raphson Algorithm (full sample case) - parameter \eqn{\theta}
}
\keyword{internal}
