% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton.R
\name{Newton_k_step}
\alias{Newton_k_step}
\title{Newton-Raphson Algorithm - k-th step of algorithm}
\usage{
Newton_k_step(theta_before_k, lambda_before_k, Hess_before_k, grad_before_k)
}
\arguments{
\item{theta_before_k}{(k-1)-th estimatation of parameter theta (shape parameter)}

\item{lambda_before_k}{(k-1)-th estimatation of parameter theta (scale parameter)}

\item{Hess_before_k}{..}

\item{grad_before_k}{..}
}
\description{
Newton-Raphson Algorithm - k-th step of algorithm
}
\keyword{internal}
