% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{sort_Y_vals}
\alias{sort_Y_vals}
\title{Sorting transformed data}
\usage{
sort_Y_vals(Y_vec_dt)
}
\arguments{
\item{Y_vec_dt}{a data table which contains value of transformed data statistic Y and
information of observation's censoring}
}
\description{
Sorting transformed data
}
\keyword{internal}
