% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{TransformedDataToY}
\alias{TransformedDataToY}
\title{Statistics of new tests}
\usage{
TransformedDataToY(T_vec_dt, theta_hat, lambda_hat)
}
\arguments{
\item{T_vec_dt}{a data table which contains value of T statistic and information of observation's censoring}

\item{theta_hat}{a estimator of parameter theta (shape parameter)}

\item{lambda_hat}{a estimator of parameter lambda (scale parameter)}
}
\description{
Statistics of new tests
}
\keyword{internal}
