% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton.R
\name{thetha_hat_approx}
\alias{thetha_hat_approx}
\title{Starting Points - theta}
\usage{
thetha_hat_approx(
  theta_before_k,
  lambda_before_k,
  Hess_before_k,
  grad_before_k
)
}
\arguments{
\item{n}{integer, number of observations in data}

\item{T_vals}{vector of observations}
}
\description{
Starting Points - theta
}
\keyword{internal}
