% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transform_to_Y}
\alias{transform_to_Y}
\title{Transformed Data to variable from extreme value distribution}
\usage{
transform_to_Y(T_vec_dt, theta_hat, lambda_hat)
}
\arguments{
\item{T_vec_dt}{a data table which contains value of T statistic and information of observation's censoring}

\item{theta_hat}{a estimator of parameter theta (shape parameter)}

\item{lambda_hat}{a estimator of parameter lambda (scale parameter)}
}
\description{
Transformed Data to variable from extreme value distribution
}
\keyword{internal}
