% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistictsUsedInComparison.R
\name{CM_test}
\alias{CM_test}
\title{Cramer-von Mises test - Right_censored Sample Case}
\usage{
CM_test(Y_vec_dt, n)
}
\arguments{
\item{Y_vec_dt}{a data table which contains value of transformed data statistic Y and
information of observation's censoring}

\item{n}{integer, number of observations in data}
}
\description{
Cramer-von Mises test - Right_censored Sample Case
}
\keyword{internal}
