% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newstatistics.R
\name{Delta_k}
\alias{Delta_k}
\title{Delta statistic}
\usage{
Delta_k(k, n, sort_Y_vec_dt)
}
\arguments{
\item{k}{numeric, the value, where the estimator will be counted}

\item{n}{integer, number of observations in data}

\item{sort_Y_vec_dt}{a data table which contains value of sorted statistic Y and
information of observation's censoring}
}
\description{
Delta statistic
}
\keyword{internal}
