% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistictsUsedInComparison.R
\name{KS_test}
\alias{KS_test}
\title{Computing value of Komogorov-Smirnov test (\eqn{KS_n^*}) for right-censored samples}
\usage{
KS_test(Y_vec_dt, n)
}
\arguments{
\item{Y_vec_dt}{a data.table object, two-columns table which contains value of statistic \eqn{Y}
and information of observation's censoring (1 - uncensored, 0 - censored).}

\item{n}{integer, amount of observations in data}
}
\value{
numeric,a value of test statistic \eqn{KS_n*}
}
\description{
Classic one-sample Kolmogorov-Smirnov test (\eqn{KS_n^*}) which was modified by ... to usage for data with
presence of censoring. Because of the Kaplan-Meier estimator is standard nonparametric estimator of ...,
The test can be use for full samples. However ... The formula used in below function is from ...
}
\keyword{internal}
