% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistictsUsedInComparison.R
\name{KS_test_full}
\alias{KS_test_full}
\title{Komogorov-Smirnov test - Full Sample Case}
\usage{
KS_test_full(X_vec, theta_hat, lambda_hat, n)
}
\arguments{
\item{X_vec}{numeric, the values from uncensured sample}

\item{theta_hat}{a estimator of parameter theta (shape parameter)}

\item{lambda_hat}{a estimator of parameter lambda (scale parameter)}

\item{n}{integer, number of observations in data}
}
\description{
Komogorov-Smirnov test - Full Sample Case
}
\keyword{internal}
