% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistictsUsedInComparison.R
\name{LS_test}
\alias{LS_test}
\title{Computing value of Liao and Shimokawa test (\eqn{LS_n*}) for right-censored samples}
\usage{
LS_test(Y_vec_dt, n)
}
\arguments{
\item{Y_vec_dt}{a data.table object, two-columns table which contains value of statistic \eqn{Y}
and information of observation's censoring (1 - uncensored, 0 - censored).}

\item{n}{integer, amount of observations in data}
}
\value{
numeric,a value of test statistic \eqn{LS_n*}
}
\description{
Classic Liao and Shimokawa test (\eqn{LS_n*}) applied to data without censoring.
The formula used in below function is from ...
}
\keyword{internal}
