% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistictsUsedInComparison.R
\name{LS_test_full}
\alias{LS_test_full}
\title{Liao and Shimokawa test - Full Sample Case}
\usage{
LS_test_full(X_vec, n)
}
\arguments{
\item{X_vec}{numeric, the values from uncensured sample}

\item{n}{integer, number of observations in data}
}
\description{
Liao and Shimokawa test - Full Sample Case
}
\keyword{internal}
