% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{m_find}
\alias{m_find}
\title{Function to find}
\usage{
m_find(x, perc_of_cens)
}
\arguments{
\item{x}{a numeric vector with observations of full sample}

\item{perc_of_cens}{numeric, percentage of censorship (value from (0, 1])}
}
\description{
Function to find
}
