% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newstatistics.R
\name{S_n_1}
\alias{S_n_1}
\alias{S_n_2}
\title{New Statistics}
\usage{
S_n_1(a, n, Y_vec_dt)

S_n_2(a, n, Y_vec_dt)
}
\arguments{
\item{a}{numeric, the value of statistic's parameter}

\item{n}{integer, number of observations in data}

\item{Y_vec_dt}{a data table which contains value of transformed statistic T (Y) and
information of observation's censoring}
}
\description{
New Statistics
}
\keyword{internal}
