% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton.R
\name{Newton_result}
\alias{Newton_result}
\title{Newton-Raphson Algorithm}
\usage{
Newton_result(T_dt)
}
\arguments{
\item{T_dt}{data table with one column with values of observation, second one contains information if the observation
is censured or not}
}
\description{
Newton-Raphson Algorithm
}
