% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton.R
\name{Newton_result}
\alias{Newton_result}
\title{Maximum Likelihood Estimation of two-parameter Weibull distribution using Newton-Raphson algorithm}
\usage{
Newton_result(T_dt)
}
\arguments{
\item{T_dt}{data table with two columns. First one contains values of observation and second one - information if the observation
is censured or not}
}
\value{
a list which contains estimation of \eqn{\theta} and \eqn{\lambda} parameters
}
\description{
Newton-Raphson algorithm is numerical method applied to find estimation of roots of function. In this case method was used
to find the root of loglikelihood function for Weibull distribution which is described by the formula:
\deqn{l(\lambda, \theta|\textbf{X}) =
\log{(\theta)}\sum^n_{i=1}\delta_i -
\theta\log{(\lambda)}\sum^n_{i=1}\delta_i +
(\theta - 1)\sum^n_{i=1}\delta_i\log{(X_i)} -
\lambda^{-\theta}\sum^n_{i=1}X_i^{\theta},}
where \eqn{\delta_i} ...
The starting points of algorithm are solution of ...
}
\examples{
X <- rweibull(100, 2, 1)
C <- runif(100, 0, m_find(X, 0.1))
T_dt <- data.table("T_val" = pmin(X, C), "delta" = pmin(X, C) == X)
Newton_result(T_dt)

}
