% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton.R
\name{theta_hat_approx}
\alias{theta_hat_approx}
\title{Starting Points - theta}
\usage{
theta_hat_approx(n, T_vals)
}
\arguments{
\item{theta_before_k}{integer, number of observations in data}

\item{lambda_before_k}{vector of observations}
}
\description{
Starting Points - theta
}
\keyword{internal}
